+++ app/helpers/extend.js
@@ -1,0 +1,1 @@
+export {default, extend} from 'ember-frost-core/helpers/extend'