+++ index.js
@@ -72,6 +72,15 @@ function isObject (object) {
   return object !== null && typeof object === 'object'
 }
 
+/**
+ * Default the babel options
+ * @param {Object} options - the options for this build
+ */
+function defaultBabel (options) {
+  options.babel = options.babel || {}
+  options.babel.optional = options.babel.optional || []
+}
+
 module.exports = {
   name: 'ember-frost-core',
 
@@ -99,13 +108,14 @@ module.exports = {
 
   init: function (app) {
     this.options = this.options || {}
-    this.options.babel = this.options.babel || {}
-    this.options.babel.optional = this.options.babel.optional || []
+    defaultBabel(this.options)
 
     if (this.options.babel.optional.indexOf('es7.decorators') === -1) {
       this.options.babel.optional.push('es7.decorators')
     }
-    this._super.init && this._super.init.apply(this, arguments)
+    if (this._super.init) {
+      this._super.init.apply(this, arguments)
+    }
   },
 
   flattenIcons: function (iconNames, subDir, srcDir) {