+++ tests/integration/components/frost-text-test.js
@@ -1,10 +1,10 @@
 import {expect} from 'chai'
 import Ember from 'ember'
 const {run} = Ember
-import {$hook, initialize} from 'ember-hook'
+import {$hook} from 'ember-hook'
 import {describeComponent, it} from 'ember-mocha'
 import hbs from 'htmlbars-inline-precompile'
-import {beforeEach} from 'mocha'
+import {beforeEach, describe} from 'mocha'
 import sinon from 'sinon'
 
 describeComponent(
@@ -14,10 +14,6 @@ describeComponent(
     integration: true
   },
   function () {
-    beforeEach(function () {
-      initialize()
-    })
-
     it('renders', function () {
       this.render(hbs`
           {{frost-text}}
@@ -36,12 +32,12 @@ describeComponent(
       expect(
         this.$('input').hasClass('frost-text-input'),
         'class "frost-text-input" is set'
-      ).to.be.true
+      ).to.equal(true)
 
       expect(
         this.$('input').hasClass('left'),
         'class "left" is set'
-      ).to.be.true
+      ).to.equal(true)
 
       expect(
         this.$('.frost-text-clear'),
@@ -63,7 +59,7 @@ describeComponent(
       expect(
         this.$('input').hasClass(align),
         'class "right" is set'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     // "autocapitalize" only works in Chrome and iOS Safari Mobile
@@ -93,7 +89,7 @@ describeComponent(
     //   expect(
     //     this.$('input').prop('autocorrect'),
     //     'autocorrect attribute is set'
-    //   ).to.be.true
+    //   ).to.equal(true)
     // })
 
     it('sets autofocus property', function () {
@@ -106,7 +102,7 @@ describeComponent(
       expect(
         this.$('input').prop('autofocus'),
         'autofocus attribute is set'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     it('set disabled property', function () {
@@ -119,7 +115,7 @@ describeComponent(
       expect(
         this.$('input').prop('disabled'),
         'disabled attribute is set'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     it('sets maxlength property', function () {
@@ -183,7 +179,7 @@ describeComponent(
       expect(
         this.$('input').prop('readonly'),
         'readonly attribute is set'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     it('set required property', function () {
@@ -196,7 +192,7 @@ describeComponent(
       expect(
         this.$('input').prop('required'),
         'required attribute is set'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     it('set spellcheck property', function () {
@@ -209,7 +205,7 @@ describeComponent(
       expect(
         this.$('input').prop('spellcheck'),
         'spellcheck attribute is set'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     it('set tabindex property', function () {
@@ -254,24 +250,24 @@ describeComponent(
       expect(
         this.$('.frost-text').hasClass('is-clear-visible'),
         'class "is-clear-visible" is not set'
-      ).to.be.false
+      ).to.equal(false)
 
       expect(
         this.$('.frost-text').hasClass('is-clear-enabled'),
         'class "is-clear-enabled" is not set'
-      ).to.be.false
+      ).to.equal(false)
 
       run(() => this.$('input').val('Test').trigger('input'))
 
       expect(
         this.$('.frost-text').hasClass('is-clear-visible'),
         'class "is-clear-visible" is set'
-      ).to.be.true
+      ).to.equal(true)
 
       expect(
         this.$('.frost-text').hasClass('is-clear-enabled'),
         'class "is-clear-enabled" is set'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     it('runs clear() which clears the input value', function () {
@@ -290,22 +286,22 @@ describeComponent(
       ).to.eql('')
     })
 
-    it('hook attr grabs frost-text as expected', function () {
-      this.render(hbs`
+    describe('when setting the hook property', function () {
+      beforeEach(function () {
+        this.render(hbs`
           {{frost-text
-            hook='my-text'
+            hook='myText'
           }}
-      `)
+        `)
+      })
 
-      expect(
-        $hook('my-text-input').hasClass('frost-text-input'),
-        'input hook is set'
-      ).to.be.true
+      it('should set the clear hook', function () {
+        expect($hook('myText-clear')).to.have.class('frost-text-clear')
+      })
 
-      expect(
-        $hook('my-text-clear').hasClass('frost-text-clear'),
-        'clear hook is set'
-      ).to.be.true
+      it('should set the input hook', function () {
+        expect($hook('myText-input')).to.have.class('frost-text-input')
+      })
     })
 
     it('calls onKeyUp closure action', function () {
@@ -324,7 +320,7 @@ describeComponent(
       expect(
         externalActionSpy.called,
         'onKeyUp closure action called'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     it('calls onInput closure action', function () {
@@ -366,7 +362,7 @@ describeComponent(
       expect(
         this.$('input').prop('disabled'),
         'disabled attribute is set'
-      ).to.be.true
+      ).to.equal(true)
     })
   }
 )