+++ tests/unit/components/frost-toggle-test.js
@@ -4,14 +4,14 @@ const {run} = Ember
 import FrostEventsProxy from 'ember-frost-core/mixins/frost-events-proxy'
 import * as utils from 'ember-frost-core/utils'
 import {describeComponent} from 'ember-mocha'
-import PropTypeMixin from 'ember-prop-types'
 import {
   beforeEach,
   describe,
   it
 } from 'mocha'
 import sinon from 'sinon'
-import SpreadMixin from 'ember-spread'
+
+import Component from 'ember-frost-core/components/frost-component'
 
 describeComponent(
   'frost-toggle',
@@ -28,30 +28,26 @@ describeComponent(
       })
     })
 
-    it('includes className frost-toggle', function () {
-      expect(component.classNames).to.include('frost-toggle')
-    })
-
     it('sets default property values correctly', function () {
       expect(
         component.get('disabled'),
         'disabled: false'
-      ).to.be.false
+      ).to.equal(false)
 
       expect(
         component.get('_trueLabel'),
         '_trueLabel: true'
-      ).to.be.true
+      ).to.equal(true)
 
       expect(
         component.get('_falseLabel'),
         '_falseLabel: false'
-      ).to.be.false
+      ).to.equal(false)
 
       expect(
         component.get('hook'),
         'hook: "undefined"'
-      ).to.be.undefined
+      ).to.equal(undefined)
 
       expect(
         component.get('size'),
@@ -90,21 +86,18 @@ describeComponent(
       ).to.eql(_isToggledDependentKeys)
     })
 
-    it('has the expected Mixins', function () {
+    it('extends the commone frost component', function () {
       expect(
-        PropTypeMixin.detect(component),
-        'PropTypeMixin Mixin is present'
-      ).to.be.true
+        component instanceof Component,
+        'is instance of Frost Component'
+      ).to.equal(true)
+    })
 
+    it('has the expected Mixins', function () {
       expect(
         FrostEventsProxy.detect(component),
         'FrostEventsProxy Mixin is present'
-      ).to.be.true
-
-      expect(
-        SpreadMixin.detect(component),
-        'SpreadMixin Mixin is present'
-      ).to.be.true
+      ).to.equal(true)
     })
 
     describe('_preferBoolean()', function () {
@@ -112,14 +105,14 @@ describeComponent(
         expect(
           component._preferBoolean('true'),
           'returned boolean true'
-        ).to.be.true
+        ).to.equal(true)
       })
 
       it('returns boolean false when passed string "false"', function () {
         expect(
           component._preferBoolean('false'),
           'returned boolean false'
-        ).to.be.false
+        ).to.equal(false)
       })
 
       it('returns value passed when not string "false" or "true"', function () {
@@ -180,7 +173,7 @@ describeComponent(
         expect(
           component.get('_isToggled'),
           '_isToggled returns true'
-        ).to.be.true
+        ).to.equal(true)
       })
 
       it('returns false when value is toggled off', function () {
@@ -192,7 +185,7 @@ describeComponent(
         expect(
           component.get('_isToggled'),
           '_isToggled returns false'
-        ).to.be.false
+        ).to.equal(false)
       })
     })
 
@@ -221,7 +214,7 @@ describeComponent(
         expect(
           cloneEventStub.called,
           'cloneEvent() method is called'
-        ).to.be.true
+        ).to.equal(true)
 
         utils.cloneEvent.restore()
       })