+++ tests/unit/components/hookable-input-test.js
@@ -1,0 +1,23 @@
+/**
+ * Unit test for the hookable-input component
+ */
+
+import {expect} from 'chai'
+import {HookMixin} from 'ember-hook'
+import {describeComponent, it} from 'ember-mocha'
+
+import {beforeEach} from 'mocha'
+
+import {unit} from 'dummy/tests/helpers/ember-test-utils/describe-component'
+
+describeComponent(...unit('hookable-input'), function () {
+  let component
+
+  beforeEach(function () {
+    component = this.subject()
+  })
+
+  it('should have the HookMixin', function () {
+    expect(HookMixin.detect(component)).to.equal(true)
+  })
+})